requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.13.4
cross-web==0.4.1
distro==1.9.0
Django==6.0.3
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.29.0
graphql-core==3.2.8
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
iniconfig==2.3.0
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
msgpack==1.1.2
openai==2.26.0
opentelemetry-api==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-django==0.61b0
opentelemetry-instrumentation-wsgi==0.61b0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg2==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sentry-sdk==2.54.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
strawberry-graphql==0.311.0
strawberry-graphql-django==0.80.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
wrapt==1.17.3
zipp==3.23.0



